Update initial build process
