+++ emhttp/plugins/dynamix/event/driver_loaded/SysDriversBuild
@@ -1,0 +1,3 @@
+#!/bin/bash
+/usr/bin/logger "Submitting SysDrivers Build" 
+/usr/local/emhttp/plugins/dynamix/include/SysDriversInit.php &> /dev/null &
